+++ frontend/components/sections/donations/5x1000/Section_4.vue
@@ -1,54 +1,23 @@
 <template>
-  <section id="fivepermille-section-4" class="padding-level-3 accent">
+  <section id="fivepermille-section-4" class="padding-level-3">
 
     <div class="section-title-center text-center">
       <h1 class="section-title">{{data ? data.title : ''}}</h1>
     </div>
 
-    <div class="text-content text-center">
-      <h3>{{data ? data.subtitle : ''}}</h3>
-    </div>
-
-    <div class="text-center btn-container">
-
-      <!--      Whatsapp-->
-      <v-btn text
-             v-ripple="{class: 'no-effects-on-btn-click'}"
-             v-if="buttons(0) && buttons(0).active"
-             class="custom-btn"
-             id="custom-whatsapp-btn"
-             data-action="share/whatsapp/share"
-             :target="buttons(0).target"
-             :href="buttons(0).href">
-      </v-btn>
-      <!--       Email-->
-      <v-btn text
-             v-ripple="{class: 'no-effects-on-btn-click'}"
-             v-if="buttons(1) && buttons(1).active"
-             class="custom-btn"
-             id="custom-email-btn"
-             :target="buttons(1).target"
-             :href="buttons(1).href">
-      </v-btn>
-      <!--       Facebook-->
-      <v-btn text
-             v-ripple="{class: 'no-effects-on-btn-click'}"
-             v-if="buttons(2) && buttons(2).active"
-             class="custom-btn"
-             id="custom-facebook-btn"
-             :target="buttons(2).target"
-             :href="buttons(2).href">
-      </v-btn>
-      <!--       Linkedin-->
-      <v-btn text
-             v-ripple="{class: 'no-effects-on-btn-click'}"
-             v-if="buttons(3) && buttons(3).active"
-             class="custom-btn"
-             id="custom-linkedin-btn"
-             :target="buttons(3).target"
-             :href="buttons(3).href">
-      </v-btn>
-
+    <div class="text-center">
+      <div>
+        <span class="text-subtitle input-title d-block" v-html="baseText ? baseText.title : ''"/>
+        <input type="number" step="0.01" class="text-center input-text" placeholder="_______"
+               v-model="ral" v-on:input="calc5x1000"/>
+      </div>
+      <div class="text-title result text-center" v-show="five_x_1000">
+        <span>{{baseText ? baseText.subtitle : ''}}</span>
+        <span class="result-number">{{five_x_1000}}</span>€
+      </div>
+      <div class="reward">
+        <span class="text-title reward-content" v-html="reward(five_x_1000)"/>
+      </div>
     </div>
   </section>
 </template>
@@ -59,17 +28,60 @@
     props: {
       data: Object
     },
+    data: () => ({
+      ranges: [
+        {index: 1, ral: 0, coeff: 0.23},
+        {index: 2, ral: 15000, coeff: 0.27},
+        {index: 3, ral: 28000, coeff: 0.38},
+        {index: 4, ral: 55000, coeff: 0.41},
+        {index: 5, ral: 75000, coeff: 0.43}
+      ],
+      ral: '',
+      irpef: 0,
+      five_x_1000: ''
+    }),
     head() {
       return {
-        meta: this.data ? this.data.meta_tags : null
+        meta: this.data.meta_tags
+      }
+    },
+    computed: {
+      baseText() {
+        return this.data ? this.data.texts.find(t => t.order === 0) : null
       }
     },
     methods: {
-      buttons(index) {
-        return this.data.buttons.find(b => b.order === index)
+      calc5x1000() {
+        this.irpef = 0
+        this.five_x_1000 = ''
+        if (this.ral > 0) {
+          this.ranges.forEach(
+            range => this.irpef += (
+              range.coeff * Math.max(0, Math.min(this.ral, this.nextRal(range.index)) - range.ral)
+            )
+          )
+          this.five_x_1000 = (this.irpef / 1000 * 5).toFixed(2)
+        }
       },
-      text(index) {
-        return this.data.texts.find(t => t.order === index)
+      nextRal(index) {
+        let next_ral = this.ranges.find(r => r.index === index + 1)
+        if (next_ral) {
+          return next_ral.ral
+        }
+        return Infinity
+      },
+      reward(donation) {
+        let reward_msg = ''
+        let texts = this.data.texts.filter(t => t.order > 0)
+        let sup = texts.filter(t => parseFloat(t.title) > donation)
+        let max = sup.length > 0 ? Math.min.apply(Math, sup.map(function(o) { return parseFloat(o.title); })) : Infinity
+        let inf = texts.filter(t => parseFloat(t.title) <= donation)
+        let min = inf.length > 0 ? Math.max.apply(Math, inf.map(function(o) { return parseFloat(o.title); })) : 0
+        let reward = texts.find(t => parseFloat(t.title) >= min && parseFloat(t.title) < max)
+        if (reward) {
+          reward_msg = reward.content
+        }
+        return reward_msg
       }
     }
   }
@@ -78,76 +90,64 @@
 <style lang="scss">
   #fivepermille-section-4 {
     margin-bottom: 10vh;
-    .section-title-center {
-      padding-top: 2vh;
-    }
     .section-title {
       margin-bottom: 5vh;
-      background-image: unset;
-      color: white;
     }
-    .text-content {
-      color: white;
-      font-family: "Crossten SemiBold";
+    .input-title {
+      margin-right: 1vw;
     }
-    .btn-container {
-      padding-top: 3vh;
-      padding-bottom: 3vh;
+    .input-text {
+      background-color: white;
+      border: 2px solid black;
+      height: 60px;
+      margin-top: 1vh;
+      margin-bottom: 2vh;
+      font-family: "Crossten Book";
+      font-size: 25px;
     }
-    .no-effects-on-btn-click {
-      color: transparent !important;
+    .result-number {
+      color: $error-color;
     }
-    .custom-btn:hover:before, .custom-btn:focus:before {
-        background-color: unset !important;
+    .reward {
+      color: $error-color;
+      margin-top: 2vh;
+      width: 100%;
     }
-    .custom-btn {
-      height: 100% !important;
-      width: auto !important;
-      .v-btn__content {
-        background-size: contain !important;
-        height: $big-button-height !important;
-        width: $big-button-height !important;
-        min-width: unset !important;
-      }
+    .reward-content {
+      max-width: 600px;
+      display: inline-block;
     }
-    #custom-whatsapp-btn {
-      .v-btn__content {
-        background: url('~assets/icons/custom-whatsapp-icon.png') center center no-repeat;
-      }
+    /* Chrome, Safari, Edge, Opera */
+    input::-webkit-outer-spin-button,
+    input::-webkit-inner-spin-button {
+      -webkit-appearance: none;
+      margin: 0;
     }
-    #custom-facebook-btn {
-      .v-btn__content {
-        background: url('~assets/icons/custom-facebook-icon.png') center center no-repeat;
-      }
-    }
-    #custom-linkedin-btn {
-      .v-btn__content {
-        background: url('~assets/icons/custom-linkedin-icon.png') center center no-repeat;
-      }
-    }
-    #custom-email-btn {
-      .v-btn__content {
-        background: url('~assets/icons/custom-email-icon.png') center center no-repeat;
-      }
+    /* Firefox */
+    input[type=number] {
+      -moz-appearance: textfield;
     }
   }
   @media #{map-get($display-breakpoints, 'sm-and-down')} {
     #fivepermille-section-4 {
+      .input-text {
+        max-width: calc(100vw - #{$md-down-logo-margin}*2);
+        height: 50px;
+        font-size: 20px;
+      }
+      .reward-content {
+        max-width: 450px;
+      }
       .text-subtitle {
         font-size: $font-size-18;
-      }
-      .custom-btn {
-        margin-bottom: 1vh;
-        margin-top: 1vh;
-        .v-btn__content {
-          height: $big-button-height * 0.7 !important;
-          width: $big-button-height * 0.7 !important;
-        }
       }
     }
   }
   @media #{map-get($display-breakpoints, 'xs-only')} {
     #fivepermille-section-4 {
+      .reward-content {
+        max-width: calc(100vw - #{$circle-icon-size}*2);
+      }
     }
   }
 </style>
